requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart

# Data processing
pandas
pyspark
kafka-python

# Resume parsing - NLP-based
pdfplumber
spacy

# Web & API clients
aiofiles
requests

# Note: After installing spacy, download the language model:
# python -m spacy download en_core_web_sm
